feat: restructure strapi schema
