+++ app/services/cms/models/formElements/StrapiDropdown.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 import { StrapiErrorRelationSchema } from "~/services/cms/models/StrapiErrorRelationSchema";
-import { HasStrapiIdSchema } from "../models/HasStrapiId";
-import { StrapiStringOptionalSchema } from "../models/StrapiStringOptional";
-import { strapiWidthLookupMap } from "../models/StrapiWidth";
+import { HasStrapiIdSchema } from "../HasStrapiId";
+import { StrapiStringOptionalSchema } from "../StrapiStringOptional";
+import { strapiWidthLookupMap } from "../StrapiWidth";
 
 export const StrapiDropdownComponentSchema = z
   .object({
