+++ app/services/cms/models/formElements/StrapiInput.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 import { StrapiErrorRelationSchema } from "~/services/cms/models/StrapiErrorRelationSchema";
-import { HasStrapiIdSchema } from "../models/HasStrapiId";
-import { StrapiStringOptionalSchema } from "../models/StrapiStringOptional";
-import { StrapiWidthSchema } from "../models/StrapiWidth";
+import { HasStrapiIdSchema } from "../HasStrapiId";
+import { StrapiStringOptionalSchema } from "../StrapiStringOptional";
+import { StrapiWidthSchema } from "../StrapiWidth";
 
 export const StrapiInputComponentSchema = z
   .object({
