+++ app/services/cms/models/formElements/StrapiSelect.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { StrapiErrorRelationSchema } from "~/services/cms/models/StrapiErrorRelationSchema";
-import { HasStrapiIdSchema } from "../models/HasStrapiId";
-import { StrapiStringOptionalSchema } from "../models/StrapiStringOptional";
+import { HasStrapiIdSchema } from "../HasStrapiId";
+import { StrapiStringOptionalSchema } from "../StrapiStringOptional";
 
 const StrapiSelectOptionSchema = z.object({
   text: z.string(),
