+++ app/services/dataListOptions/createDataListLoader.ts
@@ -2,7 +2,7 @@ import type { LoaderFunctionArgs } from "react-router";
 import { logWarning } from "~/services/logging";
 import { validateCsrfSessionFormless } from "~/services/security/csrf/validatedSession.server";
 import { getDataListOptions } from "./getDataListOptions";
-import type { DataListType } from "../cms/components/StrapiAutoSuggestInput";
+import type { DataListType } from "../cms/models/formElements/StrapiAutoSuggestInput";
 
 export function createDataListLoader(dataListType: DataListType) {
   return async function loader({ request, params }: LoaderFunctionArgs) {
