+++ app/services/dataListOptions/getDataListOptions.ts
@@ -1,7 +1,7 @@
 import { type Params } from "react-router";
 import airlines from "data/airlines/data.json";
 import airports from "data/airports/data.json";
-import type { DataListType } from "~/services/cms/components/StrapiAutoSuggestInput";
+import type { DataListType } from "~/services/cms/models/formElements/StrapiAutoSuggestInput";
 import { fetchStreetnamesForZipcode } from "~/services/gerichtsfinder/openPLZ";
 
 export type DataListOptions = {
